requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7

# Web scraping
selenium==4.18.1
beautifulsoup4==4.12.3
requests==2.31.0

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.1
pandas==2.2.1

